+++ src/main/java/com/aionemu/gameserver/ai/worlds/norsvold/Bighorn_WendigoAI2.java
@@ -15,7 +15,7 @@ public class Bighorn_WendigoAI2 extends AggressiveNpcAI2
 {
 	@Override
 	protected void handleDied() {
-		spawn(241991, getOwner().getX(), getOwner().getY(), getOwner().getZ(), (byte) getOwner().getHeading()); //Bighorn Denmaster.
+		spawn(241991, getOwner().getX(), getOwner().getY(), getOwner().getZ(), getOwner().getHeading()); //Bighorn Denmaster.
 		super.handleDied();
 		AI2Actions.deleteOwner(this);
 	}
